--no-verify

+++ inst/docs.sh
@@ -17,9 +17,7 @@ do
 done
 for f in fig2d fig3d heatmaply mcpca3d
 do
-    if [ -f vignettes/${f}.html ]; then
-       mv vignettes/${f}.html docs/articles/
-    fi
+    if [ -f vignettes/${f}.html ]; then mv vignettes/${f}.html docs/articles/; fi
 done
 
 if [ "$(uname -n | sed 's/-[0-9]*$//')" == "login-q" ]; then
